filter on all bills

+++ client/src/components/Bills components/AllBills.jsx
@@ -1,23 +1,19 @@
 import { useEffect, useState } from "react";
 import { FcAlphabeticalSortingAz, FcNumericalSorting12 } from "react-icons/fc";
-
-import Header from "../components/Header";
-import {
-  Navigate,
-  useLocation,
-  useNavigate,
-  useParams,
-} from "react-router-dom";
+import Header from "../Header";
+import { useLocation, useNavigate } from "react-router-dom";
 import { MdDeleteForever, MdEditSquare } from "react-icons/md";
 
 const AllBills = () => {
   const [data, setData] = useState({ bills: [] });
   const [categories, setcategories] = useState([]);
   const [formData, setFormData] = useState({});
+  const [searchData, setSearchData] = useState({});
+  const [search, setSearch] = useState(false);
   const [sortAmount, setSortAmount] = useState(true);
   const [sortDate, setSortDate] = useState(true);
   const [sortDescription, setSortDescription] = useState(true);
-  // const { id } = useParams();
+  const [confirmDelete, setConfirmDelete] = useState(false);
   const location = useLocation();
   const { id } = location.state;
   const navigate = useNavigate();
@@ -38,6 +34,14 @@ const AllBills = () => {
       });
   }, []);
 
+  const handleSearchChange = (e) => {
+    setSearchData({ ...searchData, [e.target.id]: e.target.value });
+    setSearch(false);
+  };
+
+  const handleSearch = (e) => {
+    setSearch(true);
+  };
   const handleChange = (e) => {
     setFormData({ categories: e.target.value });
   };
@@ -60,10 +64,10 @@ const AllBills = () => {
         .then((res) => {
           setData(res);
         });
-      }
-    };
-    
-    const handleSortingAmount = () => {
+    }
+  };
+
+  const handleSortingAmount = () => {
     setSortAmount(!sortAmount);
     setSortDescription(true);
     setSortDate(true);
@@ -86,7 +90,6 @@ const AllBills = () => {
   };
 
   const handleSortingDate = () => {
-
     setSortDate(!sortDate);
     setSortDescription(true);
     setSortAmount(true);
@@ -107,7 +110,11 @@ const AllBills = () => {
   };
 
   const handleDelete = async (index) => {
-    alert("em na thai baka delete");
+    if (!confirm("Are you sure you want to delete ")) {
+      return;
+    }
+    setConfirmDelete(true);
+
     fetch("/api/user/deleteBill/" + id, {
       method: "DELETE",
       headers: {
@@ -119,6 +126,7 @@ const AllBills = () => {
       ...prevData,
       bills: prevData.bills.filter((bill, i) => i !== index),
     }));
+    setConfirmDelete(false);
   };
 
   const formatedData = data.bills.map((d, index) => {
@@ -131,17 +139,43 @@ const AllBills = () => {
         }
       }
     }
+    if (search) {
+      if (searchData.startDate) {
+        if (searchData.startDate <= d.date) {
+        } else {
+          return;
+        }
+      }
+      if (searchData.endDate) {
+        if (searchData.endDate >= d.date) {
+        } else {
+          return;
+        }
+      }
+      if (searchData.minAmount) {
+        if (searchData.minAmount <= d.amount) {
+        } else {
+          return;
+        }
+      }
+      if (searchData.maxAmount) {
+        if (searchData.maxAmount >= d.amount) {
+        } else {
+          return;
+        }
+      }
+    }
 
     return (
       <tr
         key={index}
         className="border-b hover:bg-slate-300 hover:text-slate-700 bg-gray-100"
       >
-        <td className="p-3 px-5">{d.categories}</td>
-        <td className="p-3 px-5">{d.description}</td>
-        <td className="p-3 px-5">{d.amount}</td>
-        <td className="p-3 px-5">{d.date}</td>
-        <td className="p-3 px-5">
+        <td className="p-3 border px-5">{d.categories}</td>
+        <td className="p-3 border px-5">{d.description}</td>
+        <td className="p-3 border px-5">{d.amount}</td>
+        <td className="p-3 border px-5">{d.date}</td>
+        <td className="p-3 border px-5">
           <button
             type="button"
             className="mr-3 text-sm  py-1 px-2 rounded focus:outline-none focus:shadow-outline"
@@ -176,11 +210,61 @@ const AllBills = () => {
         <div className="p-5 flex justify-center ">
           <h1 className="text-5xl font-mono">All Bills</h1>
         </div>
+        <div className="flex flex-wrap justify-around py-3 p-2">
+          <div className="w-full sm:w-auto mb-3 sm:mb-0">
+            <label className="">Start date : </label>
+            <input
+              type="date"
+              id="startDate"
+              className="border p-3 rounded-lg w-full sm:w-auto"
+              onChange={handleSearchChange}
+            />
+          </div>
+          <div className="w-full sm:w-auto mb-3 sm:mb-0">
+            <label className="">End date : </label>
+            <input
+              type="date"
+              id="endDate"
+              className="border p-3 rounded-lg w-full sm:w-auto"
+              onChange={handleSearchChange}
+            />
+          </div>
+          <div className="w-full sm:w-auto mb-3 sm:mb-0">
+            <label className="">Minimum amount : </label>
+            <input
+              type="number"
+              id="minAmount"
+              placeholder="Minimum amount"
+              className="border p-3 rounded-lg w-full sm:w-auto"
+              onChange={handleSearchChange}
+            />
+          </div>
+          <div className="w-full sm:w-auto mb-3 sm:mb-0">
+            <label className="">Maximum amount : </label>
+            <input
+              type="number"
+              id="maxAmount"
+              placeholder="Maximum amount"
+              className="border p-3 rounded-lg w-full sm:w-auto"
+              onChange={handleSearchChange}
+            />
+          </div>
+          <div className="w-full sm:w-auto">
+            <button
+              type="button"
+              className="text-sm bg-slate-700 hover:bg-slate-500 text-white py-3 px-4 rounded"
+              onClick={handleSearch}
+            >
+              Search
+            </button>
+          </div>
+        </div>
+        {/* all bills data in table view */}
         <div className="px-3 py-4 flex justify-center">
-          <table className="w-full text-md bg-white shadow-md rounded mb-4">
+          <table className="w-full text-md border bg-white shadow-md rounded mb-4">
             <tbody>
               <tr className="border-b">
-                <th className="text-left p-3 px-5">
+                <th className="text-left border p-3 px-5">
                   <select
                     name=""
                     className="htmlForm-select border rounded-lg p-3 block w-full focus:bg-white"
@@ -200,39 +284,38 @@ const AllBills = () => {
                   </select>
                 </th>
                 <th
-                  className="text-left p-3 px-5 hover:cursor-pointer"
+                  className="text-left p-3 px-5 border hover:cursor-pointer"
                   onClick={handleSortingDescription}
                 >
                   <div className="flex items-center">
                     <div>Description</div>
                     {!sortDescription && (
                       <div className="ps-3">
-                        <FcAlphabeticalSortingAz size={"20"}/>
+                        <FcAlphabeticalSortingAz size={"20"} />
                       </div>
                     )}
                   </div>
                 </th>
                 <th
-                  className="text-left p-3 px-5 hover:cursor-pointer"
+                  className="text-left p-3 px-5 border hover:cursor-pointer"
                   onClick={handleSortingAmount}
                 >
                   <div className="flex items-center">
                     <div>Amount</div>
                     {!sortAmount && (
                       <div className="ps-3">
-                        <FcNumericalSorting12 size={"20"}/>
+                        <FcNumericalSorting12 size={"20"} />
                       </div>
                     )}
                   </div>
                 </th>
                 <th
-                  className="text-left p-3 px-5 hover:cursor-pointer"
+                  className="text-left border p-3 px-5 hover:cursor-pointer"
                   onClick={handleSortingDate}
                 >
                   date
                 </th>
                 <th className="text-left p-3 px-5">Action</th>
-                <th></th>
               </tr>
               {formatedData == "" ? (
                 <tr>
